Move 'PartialBlock' to its own module.
  Was specific to Ogmios so far, but about to re-use this for the mock chain producer.
